requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
Chroma==0.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.5
click==8.1.3
coloredlogs==15.0.1
cryptography==40.0.2
dataclasses-json==0.5.14
decorator==5.1.1
dnspython==2.4.2
elevenlabs==0.2.15
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.99.1
filelock==3.13.1
Flask==2.3.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.10.0
gpt4all==1.0.8
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.0.1
InstructorEmbedding==1.0.1
ipython==8.13.2
itsdangerous==2.1.2
gunicorn-23.0.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.316
langchainhub==0.1.13
langsmith==0.0.44
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
o==0.0.2b2
onnxruntime==1.15.1
openai==0.27.7
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
posthog==3.0.1
prompt-toolkit==3.0.38
protobuf==4.23.4
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
pymongo==4.5.0
pyOpenSSL==23.1.1
PyPika==0.48.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.30.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0
traitlets==5.9.0
transformers==4.35.0
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
